Wow so I had so many errors. Sorry, nothing prior this this compiled.
I had to change my codition in the if in my addWord moethod from this.pointer.doesntContain(curChar)   to plain old doesntContain(curChar).   I'm confused as to why I couldn't call this on method on pointer.  Maybe I need to place the method in the Node class.  I just thought of that.

I've been researching Tries and I know this is still a very weak program.